Created the Destructible script
The script adds an explosion force to a fractured gameobject on mouse click.
It takes all required parameters in the inspector to add the explosion force.
Once the fractured objects move a given distance away, the explosion force will not act upon it.

To Do:
Create the logic to shrink the parent object's collider so that the user cannot click on empty space but have an explosion force happen.


Signed-off-by: H1ddenGames <skarim3000@gmail.com>