Added materials, particle effects, updated script
The Destructible script has more options and better functionality.
Once an object is too far away, the explosion effect will no longer act upon it.
Once there are no objects within the collider of the parent object, the collider will shrink.
Once the collider has shrunk below 0.5f, the collider will be disabled.
Once a specified amount of time has passed, object fragments will be destroyed.

Added a glowing material to the object fragments to indicate to the player that it's important.

Imported VFX graph and created the explosion VFX.
At the moment it seems it's impossible to spawn the VFX or manipulate the VFX through script.

To Do: Look into decals.

Signed-off-by: H1ddenGames <skarim3000@gmail.com>