Updated destructible script
The script now uses pointer interface rather than onmouseover method.
This makes it so that the user can click as fast as they want and have the click destroy the blocks as expected.
Extracted the clean up logic from the Coroutine and put it in a normal method. The coroutine is only used for waiting for the specified time to pass.

To Fix:
There is a heavy fps penalty when there are a lot of gameobjects on the screen.
Exiting out of play mode takes a long time.

Signed-off-by: H1ddenGames <skarim3000@gmail.com>