Added hit animation, updated destructible script
The destructible script now waits until the block has been hit a specified number of times before breaking.
When the hits left variable is greater than 0, an animation will play that gives the impression of the block being hit.
There is an option called isExplosive that decides if only one random block will receive the explosion force or if all the blocks do.

Fixed the rotation snapping back to the rotation of the prefab. This was fixed by creating a parent empty GameObject and putting the prefab as the child of the empty GameObject. Put the animator on the parent and everything works properly.

Signed-off-by: H1ddenGames <skarim3000@gmail.com>